requirements.txt
mcp>=1.0.0
openai>=1.0.0
anthropic>=0.3.0
python-dotenv
pyyaml
tiktoken